+++ spec/telegram/bot/updates_controller/callback_query_context_spec.rb
@@ -1,4 +1,4 @@
-RSpec.describe Telegram::Bot::UpdatesController::CallbackQueyContext do
+RSpec.describe Telegram::Bot::UpdatesController::CallbackQueryContext do
   include_context 'telegram/bot/updates_controller'
   let(:controller_class) do
     described_class = self.described_class